+++ doc-examples/src/main/java/arez/doc/examples/at_observe2/CurrencyView.java
@@ -14,6 +14,7 @@ import javax.annotation.Nonnull;
 public abstract class CurrencyView
 {
   @CascadeDispose
+  @Nonnull
   final Currency bitcoin = new Arez_Currency();
 
   @Observe