+++ doc-examples/src/main/java/arez/doc/examples/cascade_dispose/MyService.java
@@ -5,11 +5,13 @@ import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.CascadeDispose;
 import arez.annotations.PreDispose;
+import javax.annotation.Nonnull;
 
 @ArezComponent
 public abstract class MyService
 {
   @CascadeDispose
+  @Nonnull
   final MySubService _subService = MySubService.create();
 
   @PreDispose