+++ processor/src/test/fixtures/bad_input/com/example/auto_observe/MissingNullabilityFieldAutoObserveModel.java
@@ -3,14 +3,12 @@ package com.example.auto_observe;
 import arez.annotations.ArezComponent;
 import arez.annotations.AutoObserve;
 import arez.component.ComponentObservable;
-import javax.annotation.Nullable;
 
 @ArezComponent
-abstract class ProtectedAccessFieldAutoObserveModel
+abstract class MissingNullabilityFieldAutoObserveModel
 {
   @AutoObserve
-  @Nullable
-  protected final MyType _field = null;
+  final MyType _field = null;
 
   static final class MyType
     implements ComponentObservable