+++ processor/src/test/fixtures/bad_input/com/example/cascade_dispose/MissingNullabilityFinalFieldComponent.java
@@ -1,0 +1,12 @@
+package com.example.cascade_dispose;
+
+import arez.Disposable;
+import arez.annotations.ArezComponent;
+import arez.annotations.CascadeDispose;
+
+@ArezComponent
+abstract class MissingNullabilityFinalFieldComponent
+{
+  @CascadeDispose
+  final Disposable _myElement = null;
+}