+++ processor/src/test/fixtures/input/com/example/auto_observe/RuntimeTypeValidateFieldAutoObserveModel.java
@@ -3,11 +3,13 @@ package com.example.auto_observe;
 import arez.annotations.ArezComponentLike;
 import arez.annotations.ArezComponent;
 import arez.annotations.AutoObserve;
+import javax.annotation.Nullable;
 
 @ArezComponent
 abstract class RuntimeTypeValidateFieldAutoObserveModel
 {
   @AutoObserve( validateTypeAtRuntime = true )
+  @Nullable
   final MyType _field = null;
 
   @ArezComponentLike