+++ processor/src/test/fixtures/input/com/example/component/ManagedAutoObserveArezComponentLikeReference.java
@@ -3,11 +3,13 @@ package com.example.component;
 import arez.annotations.ArezComponentLike;
 import arez.annotations.ArezComponent;
 import arez.annotations.AutoObserve;
+import javax.annotation.Nullable;
 
 @ArezComponent
 abstract class ManagedAutoObserveArezComponentLikeReference
 {
   @AutoObserve( validateTypeAtRuntime = true )
+  @Nullable
   final MyComponent _myComponent = null;
 
   @ArezComponentLike